pyproject.toml
[project]
name = "autoshot"
version = "0.1.0"
description = "Automatic window screenshot tool with similarity detection"
authors = [{name = "AutoShot Team", email = "team@example.com"}]
requires-python = ">=3.8"
dependencies = [
    "pywin32>=227",
    "Pillow>=8.0.0",
    "imagehash>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["autoshot*"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"